backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor>=3.6.0
pymongo>=4.6,<5
transformers==4.35.0
torch>=2.0.0
torchvision>=0.15.0
pillow==10.1.0
spacy==3.7.2
requests==2.31.0
python-dotenv==1.0.0
langdetect==1.0.9
python-multipart==0.0.6
aiofiles==23.2.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Additional ML Dependencies for Multimodal Fusion
timm>=0.9.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# Knowledge Graph & NER
SPARQLWrapper>=2.0.0
httpx>=0.25.0

# Enhanced Caching
diskcache>=5.6.3

# Visualization & Analytics
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional utilities
tqdm>=4.66.0
regex>=2023.10.0
